+++ src/index.template.html
@@ -86,7 +86,7 @@
         );
         height: 210.984px;
       }
-      .loading {
+      .initial-loading {
         margin: 100px auto;
         width: 200px;
         height: 50px;
@@ -94,7 +94,7 @@
         font-size: 10px;
       }
 
-      .loading > div {
+      .initial-loading > div {
         background-color: rgba(0, 0, 0, 1);
         height: 40%;
         width: 21px;
@@ -105,12 +105,12 @@
         animation: stretchdelay 1.1s infinite ease-in-out;
       }
 
-      .loading > div:first-child {
+      .initial-loading > div:first-child {
         -webkit-animation-delay: -1.2s;
         animation-delay: -1.2s;
       }
 
-      .loading > div:last-child {
+      .initial-loading > div:last-child {
         -webkit-animation-delay: -1s;
         animation-delay: -1s;
       }
@@ -144,7 +144,7 @@
 
     <noscript>
       <style>
-        .loading {
+        .initial-loading {
           display: none;
         }
 
@@ -175,7 +175,7 @@
         <div class="tmp-interpro-header"><h1>InterPro</h1></div>
         <div id="error"></div>
         <p>Loading the InterPro website, please wait.</p>
-        <div class="loading">
+        <div class="initial-loading">
           <div></div>
           <div></div>
           <div></div>
