+++ src/pages/Proteome/index.js
@@ -13,6 +13,7 @@ import Table, {
   Exporter,
   Card,
   HighlightToggler,
+  // $FlowFixMe
 } from 'components/Table';
 // $FlowFixMe
 import File from 'components/File';
