+++ src/pages/Structure/index.js
@@ -15,6 +15,7 @@ import Table, {
   PageSizeSelector,
   Exporter,
   HighlightToggler,
+  // $FlowFixMe
 } from 'components/Table';
 
 // $FlowFixMe
