+++ src/pages/Taxonomy/index.js
@@ -25,6 +25,7 @@ import Table, {
   PageSizeSelector,
   Exporter,
   HighlightToggler,
+  // $FlowFixMe
 } from 'components/Table';
 // $FlowFixMe
 import File from 'components/File';
